added some folder structure and dispatched image fetching to bkg thread
